Windows: Fixup builder\internals.go
