+++ builder/internals_windows.go
@@ -1,0 +1,8 @@
+// +build windows
+
+package builder
+
+func fixPermissions(source, destination string, uid, gid int, destExisted bool) error {
+	// chown is not supported on Windows
+	return nil
+}
